pyproject.toml
[project]
name = "portal-transparencia"
version = "0.1.0"
description = "Aplicação para verificar beneficios"
authors = [
    {name = "Pedro Senna",email = "you@example.com"}
]
requires-python = ">=3.13"
dependencies = [
    "aiofiles (>=25.1.0,<26.0.0)",
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn (>=0.42.0,<0.43.0)",
    "playwright (>=1.58.0,<2.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "rq (>=2.7.0,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "pyjwt (>=2.12.1,<3.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
